Organizing code and width scaling. Adding inheritance in attempt for portability and readability. Moving CSS from REACT imports to html imports.
