+++ static/dashboard.html
@@ -4,17 +4,8 @@
     <meta content="text/html;charset=utf-8" http-equiv="Content-Type">
     <meta content="utf-8" http-equiv="encoding">
 
-    <link rel="stylesheet" type="text/css" href="plugins/bootstrap-3.3.7-dist/css/bootstrap.css">
-    <link rel="stylesheet" type="text/css" href="static/plugins/bootstrap-3.3.7-dist/css/bootstrap.css">
-    <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='plugins/bootstrap-3.3.7-dist/css/bootstrap.css')}}">
-
-    <link rel="stylesheet" type="text/css" href="src/components/css/tabs.css">
-    <link rel="stylesheet" type="text/css" href="static/src/components/css/tabs.css">
-    <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='src/components/css/tabs.css')}}">
-
-    <link rel="stylesheet" type="text/css" href="src/components/css/header.css">
-    <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='src/components/css/header.css')}}">
-    <link rel="stylesheet" type="text/css" href="static/src/components/css/header.css">
+    <link rel="stylesheet" type="text/css" href="plugins/master.css">
+    <link rel="stylesheet" type="text/css" href="static/plugins/master.css">
 
     <style>
 
@@ -43,40 +34,6 @@
   </head>
 
   <body>
-
-<!-- Reference for Layout
-<div class="container-fluid" style="width: 100%; height: 500px; background: white;">
-<div class="row-fluid">
-    <div class="col-md-2" style="background: green;">
-        <div class="row">
-            <div class="col-md-11 col-md-offset-1" style="background: yellow; height: 300px;"></div>
-        </div>
-        <div class="row">
-            <div class="col-md-12" style="background: blue; height: 300px;"></div>
-        </div>
-        <div class="row">
-            <div class="col-md-12" style="background: orange; height: 300px;"></div>
-        </div>
-    </div>
-
-    <div class="col-md-8" style="background: black;">
-<ul class="nav nav-tabs">
-  <li role="presentation" class="active"><a href="#">Home</a></li>
-  <li role="presentation"><a href="#">Profile</a></li>
-  <li role="presentation"><a href="#">Messages</a></li>
-</ul>
- </div>
-
-    <div class="col-md-2" style="background: black;">
-        <div class="row-fluid">
-            <div class="col-md-12" style="background: yellow; height: 300px;"></div>
-            <div class="col-md-12" style="background: blue; height: 300px;"></div>
-            <div class="col-md-12" style="background: orange; height: 300px;"></div>
-        </div>
-    </div>
-</div>
-</div>
--->
 
       <div class="container-fluid" id="SystemLoadStats" style="background:; width:100%;"/>
       <!-- Direct load of the index into browser (no flask used) -->
