+++ static/plugins/css/header.css
@@ -1,0 +1,40 @@
+@font-face { font-family: MetricHPE; src: url("./fonts/MetricHPE.otf"); }
+@font-face { font-family: MetricHPE-Semi; src: url("./fonts/MetricHPE-Semi.otf"); }
+
+.navbar, .navbar-default{
+    min-height: 40px;
+    margin-bottom: 0.4%;
+}
+
+
+.navbar-brand
+{
+    position: absolute;
+    width: 100%;
+    left: 0;
+    text-align: center;
+    margin:0 auto;
+    font-size: 1.2em;
+    font-weight: bold;
+    font-family: MetricHPE;
+    padding-top: 0.8em;
+}
+
+
+.navbar-toggle {
+    z-index:3;
+}
+
+
+.navbar-left, .navbar-right {
+    font-size: 1.1em;
+    vertical-align: middle;
+    padding-top: 0.6em;
+}
+
+
+.logo{
+    padding: -2em 0 0 0;
+    width: 100px;
+    height: 20px;
+}
