+++ static/plugins/css/stats.css
@@ -1,0 +1,111 @@
+@font-face { font-family: MetricHPE; src: url("./fonts/MetricHPE.otf"); }
+@font-face { font-family: MetricHPE-Semi; src: url("./fonts/MetricHPE-Semi.otf"); }
+
+
+.statsbox{
+    background-color: #4C5E6C;
+    text-align: center;
+    min-height: 295px;
+}
+
+
+.fixedContainer{
+    height: 16em;
+}
+
+/* -- PROGRESS CIRCLE styles starts here -- */
+
+.svg-progress-circle{
+    width: 100%;
+    height: 100%;
+    display: block;
+    text-align: center;
+}
+
+
+.progress-circle{
+    stroke-width: 0.214em;
+    cx:50%;
+    cy:50%;
+    fill: none; /* keep as "none" for transparent bg. Removing this will set to default value "black" */
+}
+
+
+.progress-inactive{
+    stroke-width: 0.3em;
+    fill: none;
+    stroke:#455561;
+    cx:50%;
+    cy:50%;
+    -ms-transform: rotate(-90deg); /* IE 9 */
+    -webkit-transform: rotate(-90deg); /* Chrome, Safari, Opera */
+    -webkit-transform-origin: 50% 50%;
+    transform: rotate(-90deg); /* Firefox */
+    transform-origin: 50% 50%;
+}
+
+
+.progress-active{
+    stroke-width: 0.4em;
+    stroke: #2AD2C9;
+    border: 15px;
+    -ms-transform: rotate(-90deg); /* IE 9 */
+    -webkit-transform: rotate(-90deg); /* Chrome, Safari, Opera */
+    -webkit-transform-origin: 50% 50%;
+    transform: rotate(-90deg); /* Firefox */
+    transform-origin: 50% 50%;
+}
+
+/* progress circle styles ENDS here */
+
+
+.data-display-container{
+    margin-top:1.5em;
+}
+
+
+@media screen and (max-height: 850px){
+    .data-display-container{
+        margin-top:0.5em;
+    }
+}/* @media max-height 850 */
+
+
+.data-container-name {
+    font-size: 1.3em;
+    font-family: MetricHPE-Semi;
+    color: #FFFFFF;
+    stroke-width: 0.036em;
+    stroke: #FFFFFF;
+    text-align: center;
+}
+
+
+/* -- TEXT display styles -- */
+
+
+.progress-value{
+    font-size: 1.5em;
+    font-weight : bold;
+    font-family: MetricHPE;
+    fill: white;
+    text-anchor: middle;
+}
+
+
+.data-metrics{
+    font-size: 1.5em;
+    font-family: MetricHPE-Semi;
+    fill: #A5AEAC;
+    text-align: center;
+}
+
+
+.data-display{
+    font-size: 3.5em;
+    font-family: MetricHPE;
+    color: white;
+    text-align: center;
+    vertical-align: middle;
+    padding: 1.5em 0 0 0;
+}
