+++ static/plugins/master.css
@@ -1,0 +1,4 @@
+@import url(./bootstrap-3.3.7-dist/css/bootstrap.css);
+@import url(./css/stats.css);
+@import url(./css/header.css);
+@import url(./css/tabs.css);
