--- static/src/components/css/stats.css
@@ -1,102 +1,0 @@
-@font-face { font-family: MetricHPE; src: url("./fonts/MetricHPE.otf"); }
-@font-face { font-family: MetricHPE-Semi; src: url("./fonts/MetricHPE-Semi.otf"); }
-
-
-.widgetContainer{
-    background-color: #4C5E6C;
-    text-align: center;
-    min-height: 295px;
-}
-
-
-.fixedContainer{
-    height: 16em;
-}
-
-
-.svgCircle{
-    width: 100%;
-    height: 100%;
-    display: block;
-    text-align: center;
-}
-
-
-.precent_text{
-    font-size: 2.5em;
-    font-weight : bold;
-    font-family: MetricHPE;
-    fill: white;
-    text-anchor: middle;
-}
-
-
-.precent_symbol{
-    font-size: 2.2em;
-    font-family: MetricHPE-Semi;
-    fill: #A5AEAC;
-    text-align: center;
-}
-
-
-.dscContainer{
-    margin-top:1.5em;
-}
-
-
-@media screen and (max-height: 850px){
-    .dscContainer{
-        margin-top:0.5em;
-    }
-}/* max-height 850 */
-
-
-.dsc {
-    font-size: 1.3em;
-    font-family: MetricHPE-Semi;
-    color: #FFFFFF;
-    stroke-width: 0.036em;
-    stroke: #FFFFFF;
-    text-align: center;
-}
-
-
-.circle{
-    stroke-width: 0.214em;
-    cx:50%;
-    cy:50%;
-    fill: none; /* keep as "none" for transparent bg. Removing this will set to default value "black" */
-}
-
-
-.progress{
-    stroke-width: 0.3em;
-    fill: none;
-    stroke:#455561;
-    cx:50%;
-    cy:50%;
-    -ms-transform: rotate(-90deg); /* IE 9 */
-    -webkit-transform: rotate(-90deg); /* Chrome, Safari, Opera */
-    -webkit-transform-origin: 50% 50%;
-    transform: rotate(-90deg); /* Firefox */
-    transform-origin: 50% 50%;
-}
-
-.progress_inactive{
-    stroke: #2AD2C9;
-    -ms-transform: rotate(-90deg); /* IE 9 */
-    -webkit-transform: rotate(-90deg); /* Chrome, Safari, Opera */
-    -webkit-transform-origin: 50% 50%;
-    transform: rotate(-90deg); /* Firefox */
-    transform-origin: 50% 50%;
-}
-
-
-.main_number{
-    font-size: 3.5em;
-    font-family: MetricHPE;
-    color: white;
-    text-align: center;
-    vertical-align: middle;
-    padding: 1.5em 0 0 0;
-}
